requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cloud-init==23.4.4
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==3.4.8
DateTime==4.5
dbus-python==1.2.18
discord==1.7.3
discord.py==1.7.3
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dsc.py==1.1.3
execnet==1.9.0
ffmpeg-python==0.2.0
filelock==3.6.0
future==0.18.2
gpg==1.16.0
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.3.6
MarkupSafe==2.0.1
mock==4.0.3
more-itertools==8.10.0
multidict==6.0.2
NBT==1.5.1
netifaces==0.11.0
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
path==16.4.0
path.py==12.5.0
pexpect==4.8.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
py-cord==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycord==0.1.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==7.1.2
pytest-shutil==1.7.0
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.20.0
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.5.6
ssh-import-id==5.11
systemd-python==234
termcolor==1.1.0
texttable==1.6.4
tomli==2.0.1
ttystatus==0.38
Twisted==22.1.0
typing_extensions==4.3.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
websocket-client==1.2.3
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
zope.interface==5.4.0
